requirements.txt
streamlit
pandas
matplotlib
openai
langchain
python-dotenv
langchain_openai
plotly
xlsxwriter
